+++ shpsync/gui/resources.py
