+++ shpsync/gui/resources.qrc
