+++ shpsync/gui/shpsync_dialog.py
