+++ shpsync/gui/shpsync_dialog_base.ui
