+++ shpsync/metadata.txt
