+++ shpsync/qgis_utils.py
